requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pytest-django>=4.5.2
pytest>=7.4.0
django-filter>=23.0
